assignment4/requirements.txt
google-genai>=0.6.0
python-dotenv>=1.0.1
pydantic>=2.7.0
langchain>=0.2.0
langchain-google-genai>=1.0.0

